+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 1.2.0 - 2019-03-06
 
+### Adds
+
+- Adds: allows users to device an array of custom attributes to prevent Pjax event listeners from firing [#39](https://github.com/Pageworks/fuel-pjax/issues/39)
+
 ### Fixes
 
 - Fixes: updates IE check to use `fuel-device-manager` package